pyproject.toml
[project]
name = "emotiondetect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "dashscope>=1.23.9",
    "deepface>=0.0.93",
    "dotenv>=0.9.9",
    "faster-whisper>=1.2.0",
    "google-genai>=1.29.0",
    "idna>=3.10",
    "insightface>=0.7.3",
    "onnxruntime>=1.22.1",
    "openai>=1.97.0",
    "opencv-python>=4.12.0.88",
    "psutil>=7.0.0",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "tf-keras>=2.19.0",
    "websocket-server>=0.6.4",
    "websockets>=15.0.1",
]

[tool.uv]
required-environments = ["sys_platform == 'win32'"]
cache-dir = "D:/installs/.uv_cache"


